Try scrape Hub*
But failed by running into the problem to deal with MS-MWBF SSO issue:

The page after `fresh_login` is a page with a form that contains some
form of the credentials according to MS-MWBF, which is auto submited in
the browser by javascript. However, the scraping framework I use,
Mechanize, doesn't run javascript, which doesn't trigger the auto
redirect. I couldnt' figuer out a way to correctly simulate the js.